move HttpClient metrics to public API
