+++ ratpack-core/src/main/java/ratpack/http/client/HttpClientStats.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,15 @@
  * limitations under the License.
  */
 
-package ratpack.http.client.internal;
+package ratpack.http.client;
 
 import java.util.Map;
 
+/**
+ * Http client metrics.
+ *
+ * @since 1.9
+ */
 public class HttpClientStats {
 
   private final Map<String, HostStats> statsPerHost;
