+++ ratpack-core/src/main/java/ratpack/http/client/internal/DefaultHttpClient.java
@@ -261,6 +261,7 @@ public class DefaultHttpClient implements HttpClientInternal {
       );
   }
 
+  @Override
   public HttpClientStats getHttpClientStats() {
     return new HttpClientStats(
       hostStats.entrySet().stream().collect(Collectors.toMap(
